% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{Data frame with two numeric columns: weights and values}

\item{W}{Total weight of the knapsack}

\item{parallel}{logical parameter for parallel computation}
}
\value{
List with numbers of items that can be put into knapsack and maximum obtained value
}
\description{
This algorithm estimates how to put available items into the knapsack
in order to get the maximum possible value less than weight of knapsack.
Here brute force approach is implemented that means all possible
combinations of items O(2^n) are considered.
}
