% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Data frame with two numeric columns: weights and values}

\item{W}{Total weight of the knapsack}
}
\value{
List with numbers of items that can be put into knapsack and approximate maximum obtained value
}
\description{
This is a greedy approach to knapsack problem. Here unit value per weight is calculated for each item.
After calculating unit per weight, the items with maximum values are selected in an order.
Total value of all selected items must be below knapsack weight \code{W}.
This approach does not give an exact value but returns at least 50\% of the true maximum value.
It reduces computational complexity due to O(n*log n).
}
