% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\docType{class}
\name{knapsack_dynamic-class}
\alias{knapsack_dynamic-class}
\alias{knapsack_dynamic}
\title{Knapsack dynamic}
\arguments{
\item{df}{data frame}

\item{W}{weight}
}
\description{
Knapsack dynamic
}

