% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab06.R
\docType{package}
\name{lab06}
\alias{lab06}
\title{Knapsack}
\description{
This package includes three functions for solving Knapsack Problem.
Each function uses different approach and performs with different types of
computational complexity.
}
\details{
Functions \code{brute_force_knapsack()}, \code{greedy_knapsack()} and \code{knapsack_dynamic()} are included.
}
\seealso{
Useful links:

\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\author{
- Farid Musayev   (gitHub:"faridmusayev", liuid:farmu437)
- Mucahit Sahin   (gitHub:"mchtsahin", liuid:mucsa806)
- Selen Karaduman (gitHub:"selenkaraduman", liuid:selka266)
}
