% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack Algorithm - Brute Force}
\usage{
brute_force_knapsack(x, W, p = FALSE)
}
\arguments{
\item{x}{data.frame with column names 'v' & 'w'}

\item{W}{Knapsack Maximum Weight}

\item{p}{Bool value to pass to active parallel}
}
\value{
Return best Knapsack combination with maximum value
}
\description{
Brute Force Approch to Solve Knapsack Problem
}
