% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack class}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{an object of class data.frame with two variables v (values) and w (weights).}

\item{W}{numeric scalar object that represents the knapsack size.

In greedy approach to knapsack problem.The unit value per weight is calculated for each item.
After calculating unit per weight, the items with maximum values are selected in order.
Selected items total must be below knapsack weight \code{W}.
This approach suppose not to give an exact value but return at least 50\% of the maximum value.}
}
\description{
Greedy Knapsack class
}
