% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\docType{class}
\name{knapsack_dynamic-class}
\alias{knapsack_dynamic-class}
\alias{knapsack_dynamic}
\title{Knapsack dynamic
\code{knapsack_dynamic} uses the dynamic programming.}
\arguments{
\item{df}{an object of class data.frame with two variables v and w.}

\item{W}{numeric scalar object that represents the knapsack size.}
}
\description{
Knapsack dynamic
\code{knapsack_dynamic} uses the dynamic programming.
}

