% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\docType{class}
\name{knapsack_dynamic-class}
\alias{knapsack_dynamic-class}
\alias{knapsack_dynamic}
\title{Dynamic Knapsack}
\arguments{
\item{x}{Data frame weights and values}

\item{W}{Total weight of the knapsack}
}
\value{
maximum obtained value
}
\description{
This is a dynamic programming (DP) approach to solving knapsack problem.
Here a tabulation method is used. DP approach results into better performance
}

